Rename walmart.js -> on.js and flattened the tree
