+++ configurations/es6-test.js
@@ -1,7 +1,7 @@
 "use strict";
 
 module.exports = {
-  "extends": "defaults/configurations/walmart/es6",
+  "extends": "defaults/configurations/es6",
   "env": {
     "mocha": true,
     "phantomjs": true
