+++ package.json
@@ -23,7 +23,7 @@
     "eslint-plugin-react": "3.12.0"
   },
 
-  "main": "configurations/walmart/es6.js",
+  "main": "configurations//es6.js",
   "scripts": {
     "test": "eslint configurations rules"
   },
