+++ rules/eslint/strict/on.js
