+++ rules/react/on.js
